Move tenant logs logic to a new file
